Add scrapers and save screenshots of progress.
Scrapers names and their screenshots are named the same, so the client can find the images to show.

Also, cleaned the code a bit, moved some code into a `Browser` class
more towards best practice way to use `.then` and indenting of it.
Only do debug and trace if enabled.