Show the screenshots from the logins steps and scrapers.
Finish the `onShowScreenshots`.
Save screenshots in the `build` folder, because the static file server is already serving files from there, so it was the simplest solution.
Close the backend browser after last scraper, to save memory (it typically causes an error, if you have monitorScreenshots on, but it can be ignored for now, as it doesn't affect functionality).
Hide the monitorScreenshots after the backend browser is closed.